requirements.txt
flask
flask-cors
werkzeug
pymongo[srv]
dnspython
python-dotenv
gunicorn